Aggiunta controllo username e collegamento db

+++ h2SimpleBackendLogin/src/main/java/com/bit/h2SimpleBackendLogin/model/UserModel.java
@@ -9,14 +9,14 @@ import jakarta.persistence.Id;
 public class UserModel {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     private String username;
     private String password;
 
     // Getters & Setters
-    public Long getId() { return id; }
-    public void setId(Long id) { this.id = id; }
+    public int getId() { return id; }
+    public void setId(int id) { this.id = id; }
 
     public String getUsername() { return username; }
     public void setUsername(String username) { this.username = username; }
